async_download_service/requirements.txt
aiohttp==3.6.*
aiofiles==0.4.0
python-dotenv==0.19.0
